% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_data.R
\name{add_new_data}
\alias{add_new_data}
\title{Check if data in input folder is OK and move to raw data folder}
\usage{
add_new_data(input, output)
}
\arguments{
\item{input}{The folder, where a folder \code{toc} is located which
contains the new files.}

\item{output}{A folder, which contains a subfolder called \code{toc}, i.e.
the usually the raw data folder, into which the files will be moved to.}
}
\value{
a \code{list} which contains the individual results for each file.
  \code{TRUE} if moved, \code{FALSE} if an error occurred. Details of the error
  are in the error files in the \code{input/toc} directory.
}
\description{
Check if data in input folder is OK and move to raw data folder
}
