% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_and_overwrite_table_in_RRD.R
\name{add_to_and_overwrite_table_in_RRD}
\alias{add_to_and_overwrite_table_in_RRD}
\title{Add files to database}
\usage{
add_to_and_overwrite_table_in_RRD(toc, dbname)
}
\arguments{
\item{toc}{the (\code{data.frame}) containing the toc data for the database}

\item{dbname}{name and path to the database in which the data should be written}
}
\description{
Add files to database
}
