% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractor_toc.R
\name{extractor_toc}
\alias{extractor_toc}
\title{Extractor toc data}
\usage{
extractor_toc(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly the names of the csv files created
}
\description{
Convert all \code{.cvs} files in \code{toc} folder to \code{data.frame} and save as \code{.rds} file.
}
\details{
This function is extracting data to be added to the database (and therefore make accessible for further analysis and forecasting)
from \code{.csv} files.
}
