% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processor_toc.R
\name{pre_processor_toc}
\alias{pre_processor_toc}
\title{Preprocessor toc data}
\usage{
pre_processor_toc(input, output)
}
\arguments{
\item{input}{directory from which to read the data}

\item{output}{directory to which to write the data}
}
\value{
invisibly \code{TRUE} when completed successful
}
\description{
Copy original text files
}
